requirements.txt
keyboard==0.13.5
openai==0.27.4
Pillow==9.5.0
Pillow==10.0.0
pytesseract==0.3.10
python-dotenv==1.0.0
rumps==0.4.0
